move to proper class
